pyproject.toml
[project]
name = "moondreamobjecttracking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eclipse-zenoh>=1.3.3",
    "moondream>=0.1.0",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=10.4.0",
    "rerun-sdk>=0.22.1",
]
